refactor(skeletonCard): updated files and file names
