+++ cardskeleton/src/components/skeleton.module.css
@@ -1,0 +1,38 @@
+.card {
+  border: 1px solid black;
+  padding: 10px;
+  margin: 10px;
+  border-radius: 5px;
+  box-shadow: 2px 2px 2px 2px gray;
+  width: 300px;
+  height: 100px;
+}
+
+.title {
+  width: 150px;
+  height: 25px;
+  margin: 5px;
+  margin-bottom: 5px;
+}
+
+.content {
+  width: 97%;
+  height: 25px;
+  margin: 5px;
+}
+
+.skeleton {
+  background: linear-gradient(90deg, #e0e0e0 25%, #f5f5f5 50%, #e0e0e0 75%);
+  background-size: 200% 200%;
+  animation: skeleton-loading 1.2s ease-in-out infinite;
+  border-radius: 10px;
+}
+
+@keyframes skeleton-loading {
+  0% {
+    background-position: 200% 0%;
+  }
+  100% {
+    background-position: -200% 0%;
+  }
+}
