--- cardskeleton/src/components/skeletonCard.jsx
@@ -1,26 +1,0 @@
-const SkeletonElement = () => {
-  return (
-    <div>
-      <div
-        style={{
-          width: "150px",
-          height: "25px",
-          margin: "10px",
-          border: "1px solid black",
-          borderRadius: "12px",
-        }}
-      ></div>
-      <div
-        style={{
-          width: "250px",
-          height: "25px",
-          margin: "10px",
-          border: "1px solid black",
-          borderRadius: "12px",
-        }}
-      ></div>
-    </div>
-  );
-};
-
-export default SkeletonElement;
